Tarea 1 check idx 1.0.31

+++ Entregas/Assignement_1/Assignement1.cu
@@ -44,12 +44,12 @@ int main(int argc, char **argv){
       printf("Matrix size: nx %d ny %d\n", nx, ny);
 
       // malloc host memory
-      float *h_A, *h_B, *h_R, *omp_R , *gpu_R;
-      h_A = (float *)malloc(nBytes);
-      h_B = (float *)malloc(nBytes);
-      h_R = (float *)malloc(nBytes);
-      omp_R = (float *)malloc(nBytes);
-      gpu_R = (float *)malloc(nBytes);
+      int *h_A, *h_B, *h_R, *omp_R , *gpu_R;
+      h_A = (int *)malloc(nBytes);
+      h_B = (int *)malloc(nBytes);
+      h_R = (int *)malloc(nBytes);
+      omp_R = (int *)malloc(nBytes);
+      gpu_R = (int *)malloc(nBytes);
 
       // initialize data at host side
       initialData(h_A, nxy);
@@ -97,7 +97,7 @@ int main(int argc, char **argv){
 
 
       // malloc device global memory
-      float *d_MatA, *d_MatB, *d_MatC;
+      int *d_MatA, *d_MatB, *d_MatC;
       SAFE_CALL(cudaMalloc((void **)&d_MatA, nBytes), "Error allocating d_MatA");
       SAFE_CALL(cudaMalloc((void **)&d_MatB, nBytes), "Error allocating d_MatB");
       SAFE_CALL(cudaMalloc((void **)&d_MatC, nBytes), "Error allocating d_MatC");
